+++ src/tooltip/index.ts
@@ -1,0 +1,1 @@
+export { ToolTip } from './Tooltip';