+++ stories/liveEdit/DetailsLive.tsx
@@ -1,0 +1,26 @@
+import React from 'react';
+import { LiveProvider, LiveEditor, LiveError, LivePreview } from 'react-live';
+import { Details } from '../../src/details/Details';
+
+const DetailsDemo = `
+  function DetailsDemo() {
+    return (
+      <Details summary="headline summary">some details</Details>
+    )
+  }
+`.trim();
+
+const DetailsLive = () => {
+  const scope = { Details };
+  return (
+    <LiveProvider code={DetailsDemo} scope={scope}>
+      <div className="container">
+        <LiveEditor className="liveEditor" aria-label="editor" />
+        <LivePreview className="livePreview" aria-label="preview" />
+      </div>
+      <LiveError className="liveError" aria-label="error" />
+    </LiveProvider>
+  );
+};
+
+export default DetailsLive;