+++ stories/liveEdit/ToolTipLive.tsx
@@ -1,0 +1,35 @@
+import React from 'react';
+import { LiveProvider, LiveEditor, LiveError, LivePreview } from 'react-live';
+import { ToolTip } from '../../src/tooltip/Tooltip';
+import './style.css';
+
+const ToolTipDemo = `
+function ToolTipDemo() {
+  return (
+   <ToolTip
+          content="Here is tooltip for bottom."
+          direction="bottom"
+          className="tooltipClass"
+          background="red"
+          color="white"
+        >
+          <b> Testing tooltip for bottom position </b>
+        </ToolTip>
+  );
+}
+`.trim();
+
+const ToolTipLive = () => {
+  const scope = { ToolTip };
+  return (
+    <LiveProvider code={ToolTipDemo} scope={scope}>
+      <div className="container">
+        <LiveEditor className="liveEditor" aria-label="editor" />
+        <LivePreview className="livePreview" aria-label="preview" />
+      </div>
+      <LiveError className="liveError" aria-label="error" />
+    </LiveProvider>
+  );
+};
+
+export default ToolTipLive;